Added custom item util
