+++ src/main/java/me/smeths/and/rhetorical/customitems/CustomItemBandage.java
@@ -1,0 +1,35 @@
+package me.smeths.and.rhetorical.customitems;
+
+import org.bukkit.Material;
+import org.bukkit.event.player.PlayerDropItemEvent;
+import org.bukkit.event.player.PlayerInteractEntityEvent;
+import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.event.player.PlayerItemConsumeEvent;
+
+public class CustomItemBandage extends ItemEvents{
+
+    @Override
+    public void onLeftCLick(PlayerInteractEvent event) {
+
+    }
+
+    @Override
+    public void onRightClick(PlayerInteractEvent event) {
+
+    }
+
+    @Override
+    public void onRightClickEntity(PlayerInteractEntityEvent event) {
+
+    }
+
+    @Override
+    public void onConsume(PlayerItemConsumeEvent event) {
+
+    }
+
+    @Override
+    public void onDrop(PlayerDropItemEvent event) {
+
+    }
+}
