+++ src/main/java/me/smeths/and/rhetorical/handlers/PacketHandler.java
@@ -1,4 +1,4 @@
-package me.smeths.and.rhetorical.Handlers;
+package me.smeths.and.rhetorical.handlers;
 
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.ChatMessageType;
