+++ src/main/java/me/smeths/and/rhetorical/itemmanagaer/ItemTypes.java
